requirements.txt
beamngpy==1.35
python-dotenv==1.0.1
pandas==2.2.2
numpy==2.0.1